Move technical function to utility file
